+++ airbyte-cdk/bulk/core/load/src/test/kotlin/io/airbyte/cdk/load/schema/TableSchemaFactoryTest.kt
@@ -11,6 +11,7 @@ import io.airbyte.cdk.load.component.ColumnType
 import io.airbyte.cdk.load.data.FieldType
 import io.airbyte.cdk.load.data.IntegerType
 import io.airbyte.cdk.load.data.StringType
+import io.airbyte.cdk.load.schema.model.StreamTableSchema
 import io.airbyte.cdk.load.schema.model.TableName
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
@@ -42,6 +43,7 @@ class TableSchemaFactoryTest {
         every { mapper.toTempTableName(finalTableName) } returns tempTableName
         every { colNameResolver.getColumnNameMapping(inputSchema.keys) } returns columnNameMapping
         every { mapper.toColumnType(any()) } returns ColumnType("test_type", false)
+        every { mapper.toFinalSchema(any()) } answers { firstArg<StreamTableSchema>() }
 
         val result = factory.make(finalTableName, inputSchema, importType)